Multiplayer Section
Use this if your hardware/emulator accepts it. I am not responsible if it does not works. But here's the plot twist: if it does not works, this means your hardware/emulator does not accept it. Support files will come soon. But also, DO NOT USE THIS ON NEW NINTENDO, SONY AND XBOX. I AM AGAINST IT AND THIS WILL BREAK THEIR TERMS OF SERVICE. USE FOR OLD CONSOLES OR EMULATORS!

Made by: OceanicBro on 13.11.025 19:06 (GMT +3)